+++ packages/@uppy/companion/test/mockserver.js
@@ -5,6 +5,12 @@ const session = require('express-session')
 var authServer = express()
 
 authServer.use(session({ secret: 'grant', resave: true, saveUninitialized: true }))
+authServer.all('*/callback', (req, res, next) => {
+  req.session.grant = {
+    response: {access_token: 'fake token'}
+  }
+  next()
+})
 authServer.all('/drive/send-token', (req, res, next) => {
   req.session.grant = {
     state: 'non-empty-value' }